# Frontend Analysis Fixes - Critical Security & Performance Improvements