--- a/claudedocs/example-pool-config.env
+++ b/claudedocs/example-pool-config.env
@@ -1,0 +1,56 @@
+# LDAP Manager Configuration with Connection Pooling
+# Copy this file to .env.local and customize for your environment
+
+# Basic LDAP Configuration
+LDAP_SERVER=ldaps://your-ldap-server.example.com:636
+LDAP_BASE_DN=dc=example,dc=com
+LDAP_IS_AD=true
+LDAP_READONLY_USER=cn=service-account,ou=service-accounts,dc=example,dc=com
+LDAP_READONLY_PASSWORD=your-service-account-password
+
+# Session Management
+PERSIST_SESSIONS=true
+SESSION_PATH=./data/sessions.db
+SESSION_DURATION=30m
+
+# LDAP Connection Pool Configuration
+# Maximum number of connections in the pool
+LDAP_POOL_MAX_CONNECTIONS=10
+
+# Minimum number of connections to maintain
+LDAP_POOL_MIN_CONNECTIONS=2
+
+# Maximum time a connection can be idle before cleanup
+LDAP_POOL_MAX_IDLE_TIME=15m
+
+# Maximum lifetime of any connection
+LDAP_POOL_MAX_LIFETIME=1h
+
+# How often to run connection health checks
+LDAP_POOL_HEALTH_CHECK_INTERVAL=30s
+
+# Timeout when acquiring a connection from the pool
+LDAP_POOL_ACQUIRE_TIMEOUT=10s
+
+# Logging
+LOG_LEVEL=info
+
+# Example configurations for different deployment sizes:
+
+# Small deployment (< 10 concurrent users)
+# LDAP_POOL_MAX_CONNECTIONS=5
+# LDAP_POOL_MIN_CONNECTIONS=1
+# LDAP_POOL_MAX_IDLE_TIME=10m
+# LDAP_POOL_ACQUIRE_TIMEOUT=5s
+
+# Medium deployment (10-50 concurrent users)
+# LDAP_POOL_MAX_CONNECTIONS=15
+# LDAP_POOL_MIN_CONNECTIONS=3
+# LDAP_POOL_MAX_IDLE_TIME=20m
+# LDAP_POOL_ACQUIRE_TIMEOUT=10s
+
+# Large deployment (> 50 concurrent users)
+# LDAP_POOL_MAX_CONNECTIONS=25
+# LDAP_POOL_MIN_CONNECTIONS=5
+# LDAP_POOL_MAX_IDLE_TIME=30m
+# LDAP_POOL_ACQUIRE_TIMEOUT=15s