--- a/internal/web/templates/base.templ
+++ b/internal/web/templates/base.templ
@@ -1,6 +1,10 @@
 package templates
 
 templ base(title string) {
+	@baseWithAssets(title, "styles.css")
+}
+
+templ baseWithAssets(title, stylesPath string) {
 	<!DOCTYPE html>
 	<html lang="en" class="h-full bg-black text-white">
 		<head>
@@ -16,8 +20,8 @@ templ base(title string) {
 			<link rel="mask-icon" href="/static/safari-pinned-tab.svg" color="#000000"/>
 			<meta name="theme-color" content="#b8e9f4"/>
 			<meta name="msapplication-TileColor" content="#b8e9f4"/>
-			<link rel="preload" href="/static/styles.css" as="style"/>
-			<link rel="stylesheet" href="/static/styles.css"/>
+			<link rel="preload" href={ "/static/" + stylesPath } as="style"/>
+			<link rel="stylesheet" href={ "/static/" + stylesPath }/>
 		</head>
 		<body class="flex h-full flex-col overflow-y-scroll">
 			{ children... }