… claudedocs/cache-optimization-summary.md
… claudedocs/code-quality-analysis.md
… claudedocs/comprehensive-frontend-analysis.md
… claudedocs/css-analysis.md
… claudedocs/css-build-optimization-guide.md
… claudedocs/frontend-fixes-implementation-summary.md
… claudedocs/frontend-optimization-results.md
… claudedocs/implementation-summary.md
… claudedocs/ldap-connection-pooling.md
… claudedocs/security-analysis-report.md
… claudedocs/security-implementation.md
… claudedocs/template-caching-performance-optimization.md
… go.sum
… internal/ldap/pool_test.go
… internal/ldap_cache/cache.go
… internal/ldap_cache/cache_benchmark_test.go
… internal/options/app.go
… internal/web/assets.go
… internal/web/computers.go
… internal/web/groups.go
… internal/web/health.go
… internal/web/server.go
… internal/web/static/manifest.json
… internal/web/static/styles.1e22ce25.css
… internal/web/template_cache.go
… internal/web/template_cache_test.go
… internal/web/templates/errors.templ
… internal/web/templates/groups.templ
… internal/web/templates/login.templ
… internal/web/templates/users.templ